+++ cmd/ttn-lw-cli/commands/root.go
@@ -188,7 +188,7 @@ func preRun(tasks ...func() error) func(cmd *cobra.Command, args []string) error
 
 		api.SetRetryMax(config.Retry.Max)
 		api.SetRetryDefaultTimeout(config.Retry.DefaultTimeout)
-		api.SetRetryEnableMetadata(config.Retry.EnableMetatada)
+		api.SetRetryEnableMetadata(config.Retry.EnableMetadata)
 		api.SetRetryJitter(config.Retry.Jitter)
 
 		if config.CA != "" {